pyproject.toml
[tool.poetry]
name = "rheapose"
version = "0.1.0"
description = "Local Attention based Neural Network for Cell Segmentation"
authors = ["Najib Ishaq <najib_ishaq@zoho.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "rheapose", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
polus_images_formats_vector_to_label = { path = "./image-tools/formats/vector-to-label-tool", develop = true }
polus_images_formats_label_to_vector = { path = "./image-tools/formats/label-to-vector-tool", develop = true }
cellpose = "^3.0.8"
omnipose = "^1.0.6"
numpy = "^1.24.4"
tqdm = "^4.66.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pre-commit = "^3.7.1"
bump2version = "^1.0.1"
scikit-image = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
